+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [1.6.2] - 2024-09-23
 ### Fixed
-- Changed version of pytube to pytubefix to fix issue with downloading videos.
+- Changed version of pytube to pytube to fix issue with downloading videos.
 - Added support for handling bot detection by YouTube.
 
 ## [1.6.1] - 2024-08-17